fixed gcc warning about size_t/src/common/dbgrid.cppi printf format spec mismatch
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@43769 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/common/dbgrid.cpp
@@ -646,7 +646,8 @@ void wxDbGridTableBase::SetValueAsBool(int row, int col, bool value)
 
 void wxDbGridTableBase::ValidateRow(int row)
 {
-    wxLogDebug(wxT("ValidateRow(%i) currently on row (%i). Array count = %i"),row,m_row,m_keys.GetCount());
+    wxLogDebug(wxT("ValidateRow(%i) currently on row (%i). Array count = %lu"),
+               row, m_row, (unsigned long)m_keys.GetCount());
 
     if (row == m_row)
          return;
